Desacoplamiento total de PdfOperation, se crean las interfaces y sus implementación por ación (unir, dividir, extraer)
