+++ SwDividePDF/SwDividePDF/Negocio/Interface/IPdfSplit.cs
@@ -6,8 +6,5 @@ namespace SwDividePDF.Negocio.Interface
     {
         void SplitPages(string sourcePdfPath, string outputPdfPath, int LengthPagexDocument);
         void SplitPages(string sourcePdfPath, string outputPdfPath, int startPage, int LengthPagexDocument);
-
-        void SplitPagesByPattern(string sourcePdfPath, Dictionary<int, int> divisorPattern, string pref, string name);
-        void SplitPagesByPattern2(string sourcePdfPath, List<FilePatternManagement> filePatterns, string name);
     }
 }
