+++ SwDividePDF/SwDividePDF/SwDividePDF.csproj
@@ -70,6 +70,9 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Negocio\Implementation\PdfPatternSplit.cs" />
+    <Compile Include="Negocio\Implementation\PdfSplit.cs" />
+    <Compile Include="Negocio\Interface\IPdfPatternSplit.cs" />
     <Compile Include="Presentación\DividePDF.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -98,11 +101,9 @@
     <Compile Include="Negocio\Implementation\PdfJoin.cs" />
     <Compile Include="Negocio\Interface\BaseClass.cs" />
     <Compile Include="Negocio\ConfigPattern.cs" />
-    <Compile Include="Negocio\ConvertToImage.cs" />
     <Compile Include="Negocio\Interface\IPdfExtract.cs" />
     <Compile Include="Negocio\Interface\IPdfSplit.cs" />
     <Compile Include="Negocio\Interface\IPdfJoin.cs" />
-    <Compile Include="Negocio\PDFOperation.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Presentación\UnePDFs.cs">
